+++ packages/next/src/server/router.ts
@@ -510,109 +510,7 @@ if (
   process.env.NEXT_RUNTIME !== 'edge' &&
   process.env.NODE_ENV !== 'production'
 ) {
-  _makeResolver = async function makeResolver(
-    dir: string,
-    nextConfig: NextConfig
-  ) {
-    // use separate require variable so that these aren't
-    // included in next-server tracing
-    const _require = globalThis['require']
-    const { default: DevServer } = _require(
-      './dev/next-dev-server.js'
-    ) as typeof import('./dev/next-dev-server')
-
-    const { default: loadCustomRoutes } = _require(
-      '../lib/load-custom-routes'
-    ) as typeof import('../lib/load-custom-routes')
-
-    const { NodeNextRequest, NodeNextResponse } = _require(
-      './base-http/node'
-    ) as typeof import('./base-http/node')
-
-    const url = _require('url') as typeof import('url')
-
-    const devServer = new DevServer({
-      dir,
-      conf: nextConfig,
-    }) as any as {
-      customRoutes: UnwrapPromise<ReturnType<typeof loadCustomRoutes>>
-      router: Router
-      generateRoutes: any
-    }
-    devServer.customRoutes = await loadCustomRoutes(nextConfig)
-    const routes = devServer.generateRoutes.bind(devServer)()
-    devServer.router = new Router(routes)
-    const routeResults = new Map<string, string>()
-
-    // @ts-expect-error internal field
-    devServer.router.catchAllRoute = {
-      match: getPathMatch('/:path*'),
-      name: 'catchall route',
-      fn: async (req, _res, _params, parsedUrl) => {
-        // clean up internal query values
-        for (const key of Object.keys(parsedUrl.query || {})) {
-          if (key.startsWith('_next')) {
-            delete parsedUrl.query[key]
-          }
-        }
-
-        routeResults.set(
-          (req as any)._initUrl,
-          url.format({
-            pathname: parsedUrl.pathname,
-            query: parsedUrl.query,
-            hash: parsedUrl.hash,
-          })
-        )
-        return { finished: true }
-      },
-    } as Route
-
-    // @ts-expect-error internal field
-    devServer.router.compiledRoutes = devServer.router.compiledRoutes.filter(
-      (route) => {
-        return (
-          route.type === 'rewrite' ||
-          route.type === 'redirect' ||
-          route.type === 'header' ||
-          route.name === 'catchall route' ||
-          route.name?.includes('check')
-        )
-      }
-    )
-
-    return async function resolveRoute(
-      _req: IncomingMessage,
-      _res: ServerResponse
-    ) {
-      const req = new NodeNextRequest(_req)
-      const res = new NodeNextResponse(_res)
-      ;(req as any)._initUrl = req.url
-
-      await devServer.router.execute.bind(devServer.router)(
-        req,
-        res,
-        url.parse(req.url!, true)
-      )
-
-      if (!res.originalResponse.headersSent) {
-        res.setHeader('x-nextjs-route-result', '1')
-        const resolvedUrl = routeResults.get((req as any)._initUrl) || req.url!
-        const routeResult: {
-          url: string
-          statusCode: number
-          headers: Record<string, undefined | number | string | string[]>
-          isRedirect?: boolean
-        } = {
-          url: resolvedUrl,
-          statusCode: 200,
-          headers: {},
-        }
-
-        res.body(JSON.stringify(routeResult)).send()
-      }
-    }
-  }
+  _makeResolver = require('./lib/route-resolver').makeResolver
 }
 
 export const makeResolver = _makeResolver
